// src/app/services/models/beneficiary-ai-suggestion.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BeneficiaryAiSuggestion {
  beneficiaryId?: number;
  beneficiaryName?: string;
  createdAt?: string;
  dismissed?: boolean;
  expiresAt?: string;
  id?: number;
  seen?: boolean;
  suggestedAmount?: number;
  suggestionText?: string;
  userId?: string;
}